// EscalafonFrontend/src/app/interfaces/v-seguridad-salud-bienestar.ts
export interface VSeguridadSaludBienestar {
    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { number } - iTipoDocId
     */
    iTipoDocId?: number;

    /**
     * @type { Date } - dtSegSalBieFechaEmision
     */
    dtSegSalBieFechaEmision?: Date;

    /**
     * @type { string } - cSegSalBieAnotaciones
     */
    cSegSalBieAnotaciones?: string;

    /**
     * @type { number } - iSegSalBieId
     */
    iSegSalBieId?: number;

    /**
     * @type { number } - iArchId
     */
    iArchId?: number;

    /**
     * @type { string } - cTipoDocNombre
     */
    cTipoDocNombre?: string;

}